04 auth

+++ controllers/contacts.js
@@ -1,12 +1,13 @@
 // Підключаємо моделс для роботи mongoose з колекцією mongoDB
+// const contacts = require("../models/contacts");
 const Contact = require("../models/contacts");
-
+// валідація
 const { contactSchema, patchSchema } = require("../schemas/contacts");
 
 // функції controllers для контактів
 const getContacts = async (req, res, next) => {
   try {
-    const contacts = await Contact.find().exec();
+    const contacts = await Contact.find({ owner: req.user.id }).exec();
     res.send(contacts);
   } catch (error) {
     next(error);
@@ -24,8 +25,13 @@ const getContactById = async (req, res, next) => {
 
     const contact = await Contact.findById(id).exec();
     if (contact === null) {
-      return res.status(404).send({ message:"Contact not found"});
+      return res.status(404).send({ message: "Contact not found" });
     }
+
+    if (contact.owner.toString() !== req.user.id.toString()) {
+      return res.status(404).json({ message: "Contact not found" });
+    }
+
     res.send(contact);
   } catch (error) {
     next(error);
@@ -33,9 +39,15 @@ const getContactById = async (req, res, next) => {
 };
 
 const postContact = async (req, res, next) => {
-  const { name, email, phone } = req.body;
+  const contact = {
+    name: req.body.name,
+    email: req.body.email,
+    phone: req.body.phone,
+    owner: req.user.id,
+  };
+
   try {
-    const validation = contactSchema.validate({ name, email, phone });
+    const validation = contactSchema.validate(contact);
     if (validation.error) {
       const errorMessage = validation.error.details
         .map((error) => error.message)
@@ -43,7 +55,7 @@ const postContact = async (req, res, next) => {
       return res.status(400).send(`Validation Error: ${errorMessage}`);
     }
 
-    const result = await Contact.create({ name, email, phone });
+    const result = await Contact.create(contact);
     res.status(201).send(result);
   } catch (error) {
     next(error);
@@ -72,9 +84,18 @@ const putContact = async (req, res, next) => {
       return res.status(400).send(`Validation Error: ${errorMessage}`);
     }
 
+    // перевірити чи id юзеру === owner контакта
+    const ownContact = await Contact.findById(id);
+    // console.log(ownContact.owner.toString());
+    // console.log(req.user.id.toString());
+    // res.send(ownContact)
+    if (ownContact.owner.toString() !== req.user.id.toString()) {
+      return res.status(404).json({ message: "Contact not found" });
+    }
+
     const result = await Contact.findByIdAndUpdate(id, contact, { new: true });
     if (result === null) {
-      return res.status(404).send({ message:"Contact not found"});
+      return res.status(404).send({ message: "Contact not found" });
     }
     res.send(result);
   } catch (error) {
@@ -90,10 +111,17 @@ const deleteContact = async (req, res, next) => {
     if (!hexIdPattern.test(id)) {
       return res.status(400).json({ message: "not valid id" });
     }
-    const result = await Contact.findByIdAndDelete(id);
-    if (result === null) {
-      return res.status(404).send({ message:"Contact not found"});
+
+    // перевірити чи id юзеру === owner контакта
+    const contact = await Contact.findById(id);
+    // console.log(contact.owner.toString());
+    // console.log(req.user.id.toString());
+    if (contact.owner.toString() !== req.user.id.toString()) {
+      return res.status(404).json({ message: "Contact not found" });
     }
+
+    const result = await Contact.findByIdAndDelete(id);
+
     res.send(result);
   } catch (error) {
     next(error);
@@ -123,6 +151,13 @@ const patchContact = async (req, res, next) => {
     if (!hexIdPattern.test(id)) {
       return res.status(400).json({ message: "not valid id" });
     }
+
+    // перевірити чи id юзеру === owner контакта
+    const ownContact = await Contact.findById(id);
+    if (ownContact.owner.toString() !== req.user.id.toString()) {
+      return res.status(404).json({ message: "Contact not found" });
+    }
+
     const results = await Contact.findByIdAndUpdate(
       id,
       { favorite },
